Код, описывающий работу с массивом
